+++ packages/std/src/math/uint256.rs
@@ -100,17 +100,93 @@ impl Uint256 {
     }
 
     /// Returns a copy of the number as big endian bytes.
-    pub fn to_be_bytes(self) -> [u8; 32] {
-        let mut result = [0u8; 32];
-        self.0.to_big_endian(&mut result);
-        result
+    pub const fn to_be_bytes(self) -> [u8; 32] {
+        let words = [
+            (self.0).0[0].to_be_bytes(),
+            (self.0).0[1].to_be_bytes(),
+            (self.0).0[2].to_be_bytes(),
+            (self.0).0[3].to_be_bytes(),
+        ];
+
+        [
+            words[3][0],
+            words[3][1],
+            words[3][2],
+            words[3][3],
+            words[3][4],
+            words[3][5],
+            words[3][6],
+            words[3][7],
+            words[2][0],
+            words[2][1],
+            words[2][2],
+            words[2][3],
+            words[2][4],
+            words[2][5],
+            words[2][6],
+            words[2][7],
+            words[1][0],
+            words[1][1],
+            words[1][2],
+            words[1][3],
+            words[1][4],
+            words[1][5],
+            words[1][6],
+            words[1][7],
+            words[0][0],
+            words[0][1],
+            words[0][2],
+            words[0][3],
+            words[0][4],
+            words[0][5],
+            words[0][6],
+            words[0][7],
+        ]
     }
 
     /// Returns a copy of the number as little endian bytes.
-    pub fn to_le_bytes(self) -> [u8; 32] {
-        let mut result = [0u8; 32];
-        self.0.to_little_endian(&mut result);
-        result
+    pub const fn to_le_bytes(self) -> [u8; 32] {
+        let words = [
+            (self.0).0[0].to_le_bytes(),
+            (self.0).0[1].to_le_bytes(),
+            (self.0).0[2].to_le_bytes(),
+            (self.0).0[3].to_le_bytes(),
+        ];
+
+        [
+            words[0][0],
+            words[0][1],
+            words[0][2],
+            words[0][3],
+            words[0][4],
+            words[0][5],
+            words[0][6],
+            words[0][7],
+            words[1][0],
+            words[1][1],
+            words[1][2],
+            words[1][3],
+            words[1][4],
+            words[1][5],
+            words[1][6],
+            words[1][7],
+            words[2][0],
+            words[2][1],
+            words[2][2],
+            words[2][3],
+            words[2][4],
+            words[2][5],
+            words[2][6],
+            words[2][7],
+            words[3][0],
+            words[3][1],
+            words[3][2],
+            words[3][3],
+            words[3][4],
+            words[3][5],
+            words[3][6],
+            words[3][7],
+        ]
     }
 
     pub fn is_zero(&self) -> bool {